requirements.txt
absl-py==0.7.1
astor==0.7.1
astroid==2.2.5
autopep8==1.4.3
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Django==2.1.7
Flask==1.0.2
gast==0.2.2
get==2018.11.19
gevent==1.4.0
greenlet==0.4.15
grpcio==1.19.0
h5py==2.9.0
idna==2.8
image==1.5.27
isort==4.3.15
itsdangerous==1.1.0
Jinja2==2.10
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.1
mccabe==0.6.1
mock==2.0.0
numpy==1.16.2
pbr==5.1.3
Pillow==5.4.1
post==2018.11.20
protobuf==3.7.0
public==2018.11.20
pycodestyle==2.5.0
pylint==2.3.1
pytz==2018.9
PyYAML==5.1
query-string==2018.11.20
request==2018.11.20
requests==2.21.0
scipy==1.2.1
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
typed-ast==1.3.1
urllib3==1.24.1
Werkzeug==0.15.1
wrapt==1.11.1
